# backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: IT Rabotyagi API
  description: >
    Упрощенная спецификация API с базовыми сценариями регистрации, авторизации,
    получения профиля и просмотра списка менторов.
  version: 0.1.0
servers:
  - url: /api/v1
    description: Базовый путь для API
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Зарегистрировать нового пользователя
      operationId: registerUser
      description: >
        Создает новую учетную запись по email, паролю и никнейму и сразу
        возвращает пару access/refresh токенов для последующих запросов.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Пользователь зарегистрирован и авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Выполнить вход по email и паролю
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Пользователь авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Обновить access токен по refresh токену
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshRequest'
      responses:
        '200':
          description: Обновленная пара токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me:
    get:
      tags: [Users]
      summary: Получить профиль текущего пользователя
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Профиль авторизованного пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/statistics:
    get:
      tags: [Users]
      summary: Получить общую статистику пользователя
      operationId: getUserStatistics
      description: Возвращает общую статистику по курсам и вопросам для текущего пользователя
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Общая статистика
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStatistics'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/statistics/courses:
    get:
      tags: [Users]
      summary: Получить детальную статистику по курсам
      operationId: getUserCourseStatistics
      description: Возвращает детальную статистику по каждому курсу, на который записан пользователь
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Статистика по курсам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCourseStatisticsList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/statistics/questions:
    get:
      tags: [Users]
      summary: Получить статистику по вопросам
      operationId: getUserQuestionStatistics
      description: Возвращает общую и детальную статистику по решенным вопросам
      security:
        - BearerAuth: []
      parameters:
        - name: courseId
          in: query
          description: Фильтр по курсу (опционально)
          schema:
            type: integer
      responses:
        '200':
          description: Статистика по вопросам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserQuestionStatistics'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/profile:
    patch:
      tags: [Users]
      summary: Обновить профиль пользователя
      operationId: updateUserProfile
      description: >
        Обновляет основные данные пользователя (username, name, email, description).
        Проверяет уникальность username и email. При обновлении username выдаются новые JWT токены.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
      responses:
        '200':
          description: Обновленный профиль (и новые токены, если username изменен)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/me/avatar:
    patch:
      tags: [Users]
      summary: Обновить аватар пользователя
      operationId: updateUserAvatar
      description: Обновляет URL аватара пользователя
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAvatarUpdateRequest'
      responses:
        '200':
          description: Обновленный аватар
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAvatarResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/password:
    post:
      tags: [Users]
      summary: Изменить пароль
      operationId: changeUserPassword
      description: >
        Изменяет пароль пользователя. Требует подтверждение старым паролем.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Пароль успешно изменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /mentors:
    get:
      tags: [Mentors]
      summary: Получить список менторов
      operationId: listMentors
      parameters:
        - name: specialization
          in: query
          description: Фильтр по специализации ментора
          schema:
            type: string
        - name: limit
          in: query
          description: Количество карточек в выдаче
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Смещение для постраничной навигации
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Список менторов и краткая информация по каждому
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorList'
  /questions:
    get:
      tags: [Questions]
      summary: Получить список всех вопросов
      operationId: listQuestions
      description: >
        Возвращает список всех вопросов с их названиями и технологиями.
        Поддерживает фильтрацию по технологии, сложности, компании и поиск по тексту.
      parameters:
        - name: search
          in: query
          description: Поиск по названию и содержанию вопроса
          schema:
            type: string
        - name: technology
          in: query
          description: Фильтр по технологии (языку программирования)
          schema:
            type: string
        - name: difficulty
          in: query
          description: Фильтр по уровню сложности
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: company
          in: query
          description: Фильтр по компании
          schema:
            type: string
        - name: limit
          in: query
          description: Количество вопросов в выдаче
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Смещение для постраничной навигации
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Список вопросов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionList'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Questions]
      summary: Создать вопрос
      operationId: createQuestion
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreateRequest'
      responses:
        '201':
          description: Вопрос создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /questions/{id}:
    get:
      tags: [Questions]
      summary: Получить вопрос по ID
      operationId: getQuestionById
      description: >
        Возвращает полную информацию о вопросе по его идентификатору,
        включая варианты ответов, правильный ответ и объяснение.
      parameters:
        - name: id
          in: path
          required: true
          description: ID вопроса
          schema:
            type: integer
      responses:
        '200':
          description: Детальная информация о вопросе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Questions]
      summary: Обновить вопрос
      operationId: updateQuestion
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID вопроса
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionUpdateRequest'
      responses:
        '200':
          description: Обновленный вопрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Questions]
      summary: Удалить вопрос
      operationId: deleteQuestion
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID вопроса
          schema:
            type: integer
      responses:
        '204':
          description: Вопрос удален
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /questions/{id}/submit:
    post:
      tags: [Questions]
      summary: Отправить ответ на вопрос
      operationId: submitQuestionAnswer
      description: >
        Принимает ответ пользователя на вопрос и возвращает результат проверки.
        Сохраняет статистику ответа для отображения в профиле пользователя.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID вопроса
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionSubmitRequest'
      responses:
        '200':
          description: Результат проверки ответа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionSubmitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses:
    get:
      tags: [Courses]
      summary: Получить список курсов
      operationId: listCourses
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Список курсов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseList'
    post:
      tags: [Courses]
      summary: Создать курс
      operationId: createCourse
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '201':
          description: Курс создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /courses/{id}:
    get:
      tags: [Courses]
      summary: Получить курс
      operationId: getCourseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Детали курса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Courses]
      summary: Обновить курс
      operationId: updateCourse
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateRequest'
      responses:
        '200':
          description: Обновленный курс
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Courses]
      summary: Удалить курс
      operationId: deleteCourse
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Курс удален
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{id}/modules:
    get:
      tags: [Courses]
      summary: Получить модули курса
      operationId: listModules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список модулей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleList'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Courses]
      summary: Добавить модуль в курс
      operationId: createModule
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleCreateRequest'
      responses:
        '201':
          description: Модуль создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{id}/modules/{moduleId}:
    get:
      tags: [Courses]
      summary: Получить модуль
      operationId: getModuleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Детали модуля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Courses]
      summary: Обновить модуль
      operationId: updateModule
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleUpdateRequest'
      responses:
        '200':
          description: Обновленный модуль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Courses]
      summary: Удалить модуль
      operationId: deleteModule
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Модуль удален
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{id}/enroll:
    post:
      tags: [Progress]
      summary: Записаться на курс
      operationId: enrollCourse
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Запись подтверждена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{id}/progress:
    get:
      tags: [Progress]
      summary: Получить прогресс по курсу
      operationId: getCourseProgress
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Прогресс по курсу
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseProgress'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{id}/modules/{moduleId}/questions:
    get:
      tags: [Courses]
      summary: Получить вопросы модуля
      operationId: listModuleQuestions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID курса
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
          description: ID модуля
      responses:
        '200':
          description: Список вопросов модуля
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModuleQuestionItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Модуль заблокирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Courses]
      summary: Добавить существующий вопрос в модуль
      operationId: addQuestionToModule
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID курса
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
          description: ID модуля
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [questionId, order]
              properties:
                questionId:
                  type: integer
                  description: ID существующего вопроса
                order:
                  type: integer
                  description: Порядок вопроса в модуле
      responses:
        '201':
          description: Вопрос добавлен в модуль
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{id}/modules/{moduleId}/questions/{questionId}:
    delete:
      tags: [Courses]
      summary: Удалить вопрос из модуля
      operationId: removeQuestionFromModule
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID курса
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
          description: ID модуля
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
          description: ID вопроса
      responses:
        '200':
          description: Вопрос удален из модуля
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /courses/{courseId}/modules/{moduleId}/complete:
    post:
      tags: [Progress]
      summary: Отметить модуль завершенным
      operationId: completeModule
      security:
        - BearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Статус завершения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleCompletion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /mentors/{id}:
    get:
      tags: [Mentors]
      summary: Получить профиль ментора
      operationId: getMentorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Профиль ментора
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorProfile'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
...
    AuthRegisterRequest:
      type: object
      required: [email, password, nickname]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        nickname:
          type: string
          minLength: 1
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken:
          type: string
          description: JWT для авторизации запросов
        refreshToken:
          type: string
          description: Токен для обновления access токена
        expiresIn:
          type: integer
          description: Время жизни access токена в секундах
    AuthRefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: Действующий refresh токен для обновления access токена
    UserProfile:
      type: object
      required: [id, email, fullName, role]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          type: string
          description: Роль пользователя в системе
        avatarUrl:
          type: string
          nullable: true
          description: URL аватара пользователя
        description:
          type: string
          nullable: true
          description: Описание профиля пользователя
...
    MentorList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MentorCard'
        total:
          type: integer
          minimum: 0
          description: Общее количество доступных менторов
...
    UserStatistics:
      type: object
      required: [coursesEnrolled, coursesCompleted, overallProgressPct, questionsStatistics]
      properties:
        coursesEnrolled:
          type: integer
          description: Количество записанных курсов
        coursesCompleted:
          type: integer
          description: Количество завершенных курсов
        overallProgressPct:
          type: number
          format: float
          description: Общий процент прогресса
        questionsStatistics:
          $ref: '#/components/schemas/QuestionStatisticsSummary'
...
    UserCourseStatisticsList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserCourseStatisticsItem'
...
    UserQuestionStatistics:
      type: object
      required: [overall, byCourse]
      properties:
        overall:
          $ref: '#/components/schemas/QuestionStatisticsSummary'
        byCourse:
          type: array
          items:
            $ref: '#/components/schemas/QuestionStatisticsByCourse'
...
    UserProfileUpdateRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 50
        name:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
          maxLength: 250
        description:
          type: string
          maxLength: 150
    UserProfileUpdateResponse:
      type: object
      required: [profile]
      properties:
        profile:
          $ref: '#/components/schemas/UserProfile'
        tokens:
          $ref: '#/components/schemas/AuthTokens'
          description: Новые токены (только если username был изменен)
    UserAvatarUpdateRequest:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl:
          type: string
          format: uri
          description: HTTP/HTTPS URL аватара пользователя
    UserAvatarResponse:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl:
          type: string
    ChangePasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
          minLength: 8
        newPassword:
          type: string
          minLength: 8
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется авторизация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Данные уже существуют и не могут быть повторно использованы
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT